Update PLAN
We could also use the facade pattern to place sliders and buttons on the soundboard. So use one pattern for both or one pattern for sliders and one pattern for buttons. Styling Buttons is much easier than styling Sliders (i.e. colors) so we could use the Decorator pattern for buttons and Facade pattern for Sliders